SQLite3: check if statement is released
At C++ layer, the type `Cursor` is like an iterator. It holds
a reference to `Statement`. At Java layer, `Statement`
is just a wrapper for C++ `Statement`, and it holds a pointer
to the C++ `Statement` object; Java `Cursor` type is as the same.

But when calling `release()` on Java `Statement`, it `delete`s
the `Statement` object in C++, causing that calls on Java
`Cursor` will produce use-after-free. It's hard to handle this
(detect if released and throw exception safely rather than
segfault) in C++, so do this in Java.